# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "QueueReservationRequest": {
        "additionalProperties": false,
        "properties": {
          "duration_seconds": {
            "exclusiveMinimum": 0.0,
            "title": "Duration Seconds",
            "type": "integer"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "owner": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Owner",
            "type": "string"
          }
        },
        "required": [
          "owner",
          "duration_seconds"
        ],
        "title": "QueueReservationRequest",
        "type": "object"
      },
      "ReservationStatus": {
        "enum": [
          "queued",
          "scheduled",
          "active",
          "released",
          "expired",
          "cancelled",
          "rejected",
          "expired_pending_operation"
        ],
        "title": "ReservationStatus",
        "type": "string"
      },
...
      "TimelineCategory": {
        "enum": [
          "reservation",
          "operation",
          "health",
          "hardware",
          "system",
          "workflow"
        ],
        "title": "TimelineCategory",
        "type": "string"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Phase 4 scoped API token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Versioned local API for reserving and controlling lab benches.",
    "title": "Lab Platform Agent API",
    "version": "0.5.0-alpha"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/benches/{bench_id}/queue": {
      "get": {
        "operationId": "list_queue_api_v1_benches__bench_id__queue_get",
        "parameters": [
          {
            "in": "path",
            "name": "bench_id",
            "required": true,
            "schema": {
              "title": "Bench Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Queue Api V1 Benches  Bench Id  Queue Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Queue"
      },
      "post": {
        "operationId": "queue_reservation_api_v1_benches__bench_id__queue_post",
        "parameters": [
          {
            "in": "path",
            "name": "bench_id",
            "required": true,
            "schema": {
              "title": "Bench Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Queue Reservation Api V1 Benches  Bench Id  Queue Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Queue Reservation"
      }
    },
...
    "/api/v1/benches/{bench_id}/timeline": {
      "get": {
        "operationId": "bench_timeline_api_v1_benches__bench_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "bench_id",
            "required": true,
            "schema": {
              "title": "Bench Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TimelineCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bench Timeline Api V1 Benches  Bench Id  Timeline Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Bench Timeline"
      }
    },
...
    "/api/v1/queue/{queue_entry_id}": {
      "delete": {
        "operationId": "cancel_queue_api_v1_queue__queue_entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "queue_entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queue Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Queue"
      }
    },
    "/api/v1/reservations": {
      "get": {
        "operationId": "list_reservations_api_v1_reservations_get",
        "parameters": [
          {
            "in": "query",
            "name": "bench_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bench Id"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReservationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "starts_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Starts After"
            }
          },
          {
            "in": "query",
            "name": "starts_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Starts Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Reservations Api V1 Reservations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Reservations"
      },
      "post": {
        "operationId": "create_reservation_api_v1_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Reservation Api V1 Reservations Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Reservation"
      }
    },
    "/api/v1/reservations/{reservation_id}": {
      "get": {
        "operationId": "get_reservation_api_v1_reservations__reservation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Reservation Api V1 Reservations  Reservation Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Reservation"
      }
    },
    "/api/v1/reservations/{reservation_id}/cancel": {
      "post": {
        "operationId": "cancel_reservation_api_v1_reservations__reservation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Reservation Api V1 Reservations  Reservation Id  Cancel Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Reservation"
      }
    },
    "/api/v1/reservations/{reservation_id}/extend": {
      "post": {
        "operationId": "extend_reservation_api_v1_reservations__reservation_id__extend_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Extend Reservation Api V1 Reservations  Reservation Id  Extend Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Extend Reservation"
      }
    },
    "/api/v1/reservations/{reservation_id}/release": {
      "post": {
        "operationId": "release_reservation_api_v1_reservations__reservation_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "reservation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reservation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Release Reservation Api V1 Reservations  Reservation Id  Release Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Release Reservation"
      }
    },
...
    "/api/v1/workflow-runs/{workflow_run_id}": {
      "get": {
        "operationId": "get_workflow_run_api_v1_workflow_runs__workflow_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Workflow Run Api V1 Workflow Runs  Workflow Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Workflow Run"
      }
    },
    "/api/v1/workflow-runs/{workflow_run_id}/cancel": {
      "post": {
        "operationId": "cancel_workflow_run_api_v1_workflow_runs__workflow_run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Workflow Run Api V1 Workflow Runs  Workflow Run Id  Cancel Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Workflow Run"
      }
    },
...
    "/api/v1/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Workflows Api V1 Workflows Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List Workflows"
      }
    },
    "/api/v1/workflows/{workflow_name}": {
      "get": {
        "operationId": "get_workflow_api_v1_workflows__workflow_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_name",
            "required": true,
            "schema": {
              "title": "Workflow Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Workflow Api V1 Workflows  Workflow Name  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Workflow"
      }
    },
...
  }
}